1. 完善 评论功能的后台逻辑

+++ src/main/java/com/seaxll/community/dto/CommentDTO.java
@@ -12,7 +12,7 @@ import lombok.Data;
  */
 @Data
 public class CommentDTO {
-    private Long parentID;
+    private Integer parentID;
     private String content;
     private Integer type;
 }
