+++ src/main/java/com/seaxll/community/model/Comment.java
@@ -14,8 +14,8 @@ import java.sql.Timestamp;
  */
 @Data
 public class Comment {
-    private Long id;
-    private Long parentId;
+    private Integer id;
+    private Integer parentId;
     private Integer type;
     private Integer commentatorId;
     private Timestamp gmtCreate;
